3-10

+++ 3_09/ex_01790.c
@@ -4,11 +4,11 @@
 int get_kth_number(int n, int k) {
     int div = 1, exp = 1;
     while (n/div != 0) {
-        if (9 * div * exp < k) {
+        if (9 * (long long)div * exp < k) {
             k -= 9 * div * exp; div *= 10; exp += 1;
         }
         else {
-            if ((n - div + 1) * exp < k) return -1;
+            if ((n - div + 1) * (long long)exp < k) return -1;
             int th1 = (k-1)/exp + div;
             int th2 = (k-1)%exp;
             th2 = exp - th2 - 1;
